requirements.txt
pysolr==3.8.1
azure-storage-blob==12.3.0
bert-extractive-summarizer==0.4.0
